SquireWorkspaceManager major refactor
Split the code from SquireWorkspaceManager into different services, similar to the setup for Google's API service wrappers (DirectoryService, etc.).
